requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
google-auth==1.10.1
gspread==3.1.0
idna==2.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.7
